+++ packages/iov-bns/types/bnsconnector.d.ts
@@ -2,4 +2,4 @@ import { ChainConnector, ChainId } from "@iov/bcp";
 /**
  * A helper to connect to a bns-based chain at a given url
  */
-export declare function bnsConnector(url: string, expectedChainId?: ChainId): ChainConnector;
+export declare function createBnsConnector(url: string, expectedChainId?: ChainId): ChainConnector;