+++ packages/iov-ethereum/types/index.d.ts
@@ -1,6 +1,6 @@
 export { Erc20ApproveTransaction, Erc20Options, Erc20TokensMap, isErc20ApproveTransaction } from "./erc20";
 export { EthereumConnection, EthereumConnectionOptions } from "./ethereumconnection";
-export { ethereumConnector } from "./ethereumconnector";
+export { createEthereumConnector } from "./ethereumconnector";
 export { ethereumCodec, EthereumCodec, EthereumCodecOptions } from "./ethereumcodec";
 export { SwapIdPrefix } from "./serialization";
 export { pubkeyToAddress, toChecksummedAddress } from "./address";